+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- **Metrics**: Support adding multiple metric values to a metric name
 - **Utilities**: Add new `Validator` utility to validate inbound events and responses using JSON Schema
 
 ## [1.5.0] - 2020-09-04